React Lifting State Lab is completed.
