+++ src/App.js
@@ -2,15 +2,13 @@ import { useState } from "react";
 import eventsData from "./data";
 import { v1 as generateUniqueID } from "uuid";
 // import Attendees from "./Attendees";
-// import Event from "./Components/Event";
+import Event from "./Components/Event";
 import Footer from "./Components/Footer";
 import Header from "./Components/Header";
 import NewEventForm from "./Components/NewEventForm";
 
 function App() {
   const [events, setEvents] = useState(eventsData);
-
-  const [showAttendees, setShowAttendees] = useState(false);
 
   const [selectOption, setSelectOption] = useState("");
 
@@ -70,10 +68,6 @@ function App() {
     setEvents([event, ...events]);
   }
 
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
-  }
-
   function updateEventAttendance(eventId, attendeeId) {
     const eventArray = [...events];
     const eventIndex = eventArray.findIndex((event) => eventId === event.id);
@@ -90,84 +84,38 @@ function App() {
 
   return (
     <div className="App">
-      <Header />
-      <Footer key={Footer} />
-      <newEventForm
-        handleSubmit={handleSubmit}
-        handleTextChange={handleTextChange}
-        newEvent={newEvent}
-        handleSelectChange={handleSelectChange}
-      />
+      <>
+        <Header />
+      </>
+
       <main>
+        <div className="new-event"></div>
+        <>
+          <NewEventForm
+            handleSubmit={handleSubmit}
+            handleTextChange={handleTextChange}
+            newEvent={newEvent}
+            handleSelectChange={handleSelectChange}
+          />
+        </>
         <div className="events">
           <ul>
             {events.map((event) => {
               const { people: attendees } = event;
 
               return (
-                <>
-                  <li key={event.id}>
-                    <img src={event.eventImage} alt={event.name} />
-                    <h5>
-                      {event.name} {event.eventType}
-                    </h5>
-                    <br />
-                    <span>Organized by: {event.organizer} </span>
-                    <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>
-                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-                      </button>
-
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee, index) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img
-                                    src={attendee.avatar}
-                                    alt={attendee.firstName}
-                                  />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button
-                                    className="clickable"
-                                    onClick={() =>
-                                      updateEventAttendance(
-                                        event.id,
-                                        attendee.id
-                                      )
-                                    }
-                                  >
-                                    Attending:
-                                  </button>
-                                  <span>
-                                    {attendee.attendance ? "✅" : "❌"}
-                                  </span>
-                                </p>
-
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
-                  </li>
-                </>
+                <Event
+                  key={event.id}
+                  event={event}
+                  attendees={attendees}
+                  updateEventAttendance={updateEventAttendance}
+                />
               );
             })}
           </ul>
         </div>
       </main>
+      <Footer key={Footer} />
     </div>
   );
 }
